# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API RESTful Usuarios - OpenAPI 3.0
  description: >-
    Esta é uma API RESTful para cadastro de usuários em um banco de dados, foi
    construída utilizando Java e Spring Framework.
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: tyerryzanca@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'
servers:
  - url: 'https://petstore3.swagger.io/api/v3'
tags:
  - name: usuarios
    description: >-
      API RESTful em Java utilizando o Spring Framework para consulta de
      usuários.
paths:
  /api/usuarios:
    post:
      tags:
        - user
      summary: Cria um novo usuário
      description: 'Cria um novo usuário baseado nos dados de Nome, Email e Senha.'
      operationId: cadastrarUsuario
      requestBody:
        description: Cria um objeto usuário
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '201':
          description: Created
        '500':
          description: Internal Server Error
    get:
      tags:
        - user
      summary: Encontra todos os usuários cadastrados
      description: ''
      operationId: consultarTodosUsuarios
      parameters: []
      responses:
        '200':
          description: OK
        '404':
          description: Users Not Found
        '500':
          description: Internal Server Error
  '/api/usuarios/{id}':
    get:
      tags:
        - user
      summary: Encontra um usuário existente pelo id
      description: ''
      operationId: consultarUsuario
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
        '404':
          description: User Not Found
        '500':
          description: Internal Server Error
    put:
      tags:
        - user
      summary: Atualiza um usuário existente pelo id
      description: ''
      operationId: atualizarUsuario
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '200':
          description: OK
        '404':
          description: User Not Found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - user
      summary: Deleta um usuário existente pelo id
      description: ''
      operationId: deletarUsuario
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
components:
  schemas:
    Usuario:
      type: object
      properties:
        nome:
          type: string
          example: João
        email:
          type: string
          example: joao@email.com
        senha:
          type: string
          example: facil123